+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 Full changelog for Bootstrap 4 changelog
 
-## v0.02 - Badge, Button and Label - 2017-11-xx
+## v0.02 - Badge, Button and Progress bar - 2017-11-xx
 
 * Added Badge view helper.
 * Added Button view helper.
-* Added Label view helper.
+* Added Progress bar view helper.
 * Updated README, examples for each view helper and callable methods. 
 
 ## v0.01 - Jumbotron - 2017-11-13